+++ capstone-project-be.Application/Features/TouristAttractions/Handles/UpdateTouristAttractionHandler.cs
@@ -1,5 +1,10 @@
 ﻿using AutoMapper;
 using capstone_project_be.Application.DTOs.Accommodations;
+using capstone_project_be.Application.DTOs.Restaurant_RestaurantCategories;
+using capstone_project_be.Application.DTOs.RestaurantPhotos;
+using capstone_project_be.Application.DTOs.Restaurants;
+using capstone_project_be.Application.DTOs.TouristAttraction_TouristAttractionCategories;
+using capstone_project_be.Application.DTOs.TouristAttractionPhotos;
 using capstone_project_be.Application.DTOs.TouristAttractions;
 using capstone_project_be.Application.Features.TouristAttractions.Requests;
 using capstone_project_be.Application.Interfaces;
@@ -12,11 +17,13 @@ namespace capstone_project_be.Application.Features.TouristAttractions.Handles
     public class UpdateTouristAttractionHandler : IRequestHandler<UpdateTouristAttractionRequest, object>
     {
         private readonly IUnitOfWork _unitOfWork;
+        private readonly IStorageRepository _storageRepository;
         private readonly IMapper _mapper;
 
-        public UpdateTouristAttractionHandler(IUnitOfWork unitOfWork, IMapper mapper)
+        public UpdateTouristAttractionHandler(IUnitOfWork unitOfWork, IStorageRepository storageRepository, IMapper mapper)
         {
             _unitOfWork = unitOfWork;
+            _storageRepository = storageRepository;
             _mapper = mapper;
         }
 
@@ -31,8 +38,10 @@ namespace capstone_project_be.Application.Features.TouristAttractions.Handles
                 };
             }
 
-            var touristAttraction = _mapper.Map<TouristAttraction>(request.TouristAttractionData);
+            var touristAttractionData = request.UpdateTouristAttractionData;
+            var touristAttraction = _mapper.Map<TouristAttraction>(touristAttractionData);
             touristAttraction.TouristAttractionId = touristAttractionId;
+            touristAttraction.CreatedAt = DateTime.Now;
 
             var cityId = touristAttraction.CityId;
             var cityList = await _unitOfWork.CityRepository.Find(c => c.CityId == cityId);
@@ -45,17 +54,67 @@ namespace capstone_project_be.Application.Features.TouristAttractions.Handles
                 };
             }
 
-            var tA_TACategoryList = await _unitOfWork.TA_TACategoryRepository.
-                Find(ta => ta.TouristAttractionId == touristAttractionId);
-            await _unitOfWork.TA_TACategoryRepository.DeleteRange(tA_TACategoryList);
-            tA_TACategoryList = touristAttraction.TouristAttraction_TouristAttractionCategories.ToList();
-            await _unitOfWork.TA_TACategoryRepository.AddRange(tA_TACategoryList);
+            var ta_TACategoryList = await _unitOfWork.TA_TACategoryRepository.Find(ta => ta.TouristAttractionId == touristAttractionId);
+            await _unitOfWork.TA_TACategoryRepository.DeleteRange(ta_TACategoryList);
+            var ta_TACategoryData = touristAttractionData.TA_TACategories;
+            if (ta_TACategoryData != null)
+            {
+                string[] parts = ta_TACategoryData.Split(',');
+                var ta_TACategories = new List<CRUDTA_TACategoryDTO>();
+                foreach (string part in parts)
+                {
+                    ta_TACategories.Add(
+                        new CRUDTA_TACategoryDTO
+                        {
+                            TouristAttractionId = touristAttractionId,
+                            TouristAttractionCategoryId = int.Parse(part)
+                        });
+                }
+                ta_TACategoryList = _mapper.Map<IEnumerable<TouristAttraction_TouristAttractionCategory>>(ta_TACategories.Distinct());
+                await _unitOfWork.TA_TACategoryRepository.AddRange(ta_TACategoryList);
+            }
 
-            var touristAttractionPhotoList = await _unitOfWork.TouristAttractionPhotoRepository.
-                Find(ta => ta.TouristAttractionId == touristAttractionId);
-            await _unitOfWork.TouristAttractionPhotoRepository.DeleteRange(touristAttractionPhotoList);
-            touristAttractionPhotoList = touristAttraction.TouristAttractionPhotos.ToList();
-            await _unitOfWork.TouristAttractionPhotoRepository.AddRange(touristAttractionPhotoList);
+            var deletePhotos = touristAttractionData.DeletePhotos;
+            if (deletePhotos != null)
+            {
+                var deletePhotoIds = new List<int>();
+                var parts = deletePhotos.Split(",");
+                foreach (string part in parts)
+                {
+                    deletePhotoIds.Add(int.Parse(part));
+                }
+                var touristAttractionPhotoList = await _unitOfWork.TouristAttractionPhotoRepository.
+                    Find(tap => tap.TouristAttractionId == touristAttractionId && deletePhotoIds.Contains(tap.TouristAttractionPhotoId));
+
+                foreach (var tap in touristAttractionPhotoList)
+                {
+                    await _storageRepository.DeleteFileAsync(tap.SavedFileName);
+                }
+                await _unitOfWork.TouristAttractionPhotoRepository.DeleteRange(touristAttractionPhotoList);
+            }
+
+            var photoData = touristAttractionData.Photos;
+            if (photoData != null)
+            {
+                var touristAttractionPhotos = new List<CRUDTouristAttractionPhotoDTO>();
+                foreach (var photo in photoData)
+                {
+                    if (photo != null)
+                    {
+                        var savedFileName = GenerateFileNameToSave(photo.FileName);
+                        touristAttractionPhotos.Add(
+                            new CRUDTouristAttractionPhotoDTO
+                            {
+                                TouristAttractionId = touristAttractionId,
+                                PhotoURL = await _storageRepository.UpLoadFileAsync(photo, savedFileName),
+                                SavedFileName = savedFileName
+                            }
+                            );
+                    }
+                }
+                var touristAttractionPhotoList = _mapper.Map<IEnumerable<TouristAttractionPhoto>>(touristAttractionPhotos);
+                await _unitOfWork.TouristAttractionPhotoRepository.AddRange(touristAttractionPhotoList);
+            }
 
             await _unitOfWork.TouristAttractionRepository.Update(touristAttraction);
             await _unitOfWork.Save();
@@ -65,6 +124,13 @@ namespace capstone_project_be.Application.Features.TouristAttractions.Handles
                 IsSuccess = true,
                 Message = "Update địa điểm du lịch giải trí thành công"
             };
+        }
+
+        private string? GenerateFileNameToSave(string incomingFileName)
+        {
+            var fileName = Path.GetFileNameWithoutExtension(incomingFileName);
+            var extension = Path.GetExtension(incomingFileName);
+            return $"{fileName}-{DateTime.Now.ToString("yyyyMMddHHmmss")}{extension}";
         }
     }
 }
